requirements.txt
fastapi>=0.95.0
uvicorn>=0.25.0
langchain-openai>=0.3.10
langchain-community>=0.0.27
langchain-core>=0.1.31
openai>=1.12.0
yfinance>=0.2.37
python-dotenv>=1.0.0
pydantic>=2.5.3